refactor: header 공통컴포넌트 검색페이지에서만 사용하도록 수정 searchHeader로 수정 및 기존 header 공통컴포넌트 사용하고 있는 페이지에서 제거 (#362)
