+++ src/pages/class/Class.tsx
@@ -2,16 +2,13 @@ import { useParams } from 'react-router-dom';
 import { useGetLessonDetail } from '@/pages/class/apis/queries';
 import { headerStyle } from '@/pages/class/class.css';
 import ClassButtonWrapper from '@/pages/class/components/ClassButtonWrapper/ClassButtonWrapper';
-import ClassHeader from '@/pages/class/components/ClassHeader/ClassHeader';
 import ClassInfoWrapper from '@/pages/class/components/ClassInfoWrapper/ClassInfoWrapper';
 import TabWrapper from '@/pages/class/components/TabWrapper/TabWrapper';
 import Error from '@/pages/error/Error';
 import Divider from '@/shared/components/Divider/Divider';
-import { useIntersectCallback } from '@/shared/hooks/useIntersectCallback';
 
 const Class = () => {
   const { id } = useParams<{ id: string }>();
-  const [targetRef, isWhite] = useIntersectCallback(false);
 
   if (!id) {
     return <Error />;
@@ -33,13 +30,10 @@ const Class = () => {
   return (
     <>
       <div
-        ref={targetRef}
         className={headerStyle}
         style={{
           backgroundImage: `url(${imageUrl})`,
-        }}>
-        <ClassHeader isWhite={isWhite} lessonName={data.name} />
-      </div>
+        }}></div>
 
       <ClassInfoWrapper lessonData={data} />
       <Divider direction="horizontal" color="gray1" length="100%" thickness="1.2rem" />
