+++ src/pages/instructor/classDetail/ClassDetail.tsx
@@ -4,7 +4,6 @@ import * as styles from '@/pages/instructor/classDetail/classDetail.css';
 import StudentCard from '@/pages/instructor/classDetail/components/StudentCard/StudentCard';
 import ClassCard from '@/shared/components/ClassCard/ClassCard';
 import Flex from '@/shared/components/Flex/Flex';
-import Header from '@/shared/components/Header/Header';
 import Text from '@/shared/components/Text/Text';
 
 const ClassDetail = () => {
@@ -24,11 +23,6 @@ const ClassDetail = () => {
 
   return (
     <div className={styles.layoutStyle}>
-      <Header.Root isColor={true}>
-        <Header.BackIcon />
-        <Header.Title title="내 클래스 관리" />
-      </Header.Root>
-
       <div className={styles.containerStyle}>
         <Flex tag="section" gap="1.6rem" direction="column" marginBottom="2.8rem">
           <Text tag="b4" color="gray9">
