--- src/pages/onboarding/components/OnboardingHeader/onboardingHeader.css.ts
@@ -1,8 +1,0 @@
-import { style } from '@vanilla-extract/css';
-import { vars } from '@/shared/styles/theme.css';
-
-export const onboardingHeaderStyle = style({
-  zIndex: 1,
-
-  backgroundColor: vars.colors.white,
-});
